Move all `isXyz` constants to new environments constants
