+++ src/backend/anticheat/utils.ts
@@ -1,4 +1,3 @@
-import { isMac, isWindows } from '../constants'
 import { logInfo, LogPrefix, logWarning } from '../logger/logger'
 import { readFileSync, writeFileSync } from 'graceful-fs'
 import { AntiCheatInfo } from 'common/types'
@@ -6,6 +5,7 @@ import { runOnceWhenOnline } from '../online_monitor'
 import { axiosClient } from 'backend/utils'
 import { join } from 'node:path'
 import { appFolder } from 'backend/constants/paths'
+import { isMac, isWindows } from 'backend/constants/environment'
 
 const anticheatDataPath = join(appFolder, 'areweanticheatyet.json')
