+++ src/backend/launcher.ts
@@ -26,14 +26,8 @@ import {
   defaultWinePrefix,
   fixesPath,
   flatPakHome,
-  isLinux,
-  isMac,
-  isWindows,
-  isSteamDeckGameMode,
   userHome,
-  publicDir,
-  isCLINoGui,
-  isIntelMac
+  publicDir
 } from './constants'
 import {
   constructAndUpdateRPC,
@@ -95,6 +89,14 @@ import { updateGOGPlaytime } from './storeManagers/gog/games'
 import { addRecentGame } from './recent_games/recent_games'
 import { tsStore } from './constants/key_value_stores'
 import { defaultUmuPath, runtimePath } from './constants/paths'
+import {
+  isCLINoGui,
+  isLinux,
+  isMac,
+  isSteamDeckGameMode,
+  isWindows,
+  isIntelMac
+} from './constants/environment'
 
 let powerDisplayId: number | null
