+++ src/backend/main.ts
@@ -61,14 +61,8 @@ import {
   userHome,
   icon,
   configPath,
-  isSteamDeckGameMode,
-  isCLIFullscreen,
-  isCLINoGui,
-  isFlatpak,
   publicDir,
-  createNecessaryFolders,
-  isWindows,
-  isMac
+  createNecessaryFolders
 } from './constants'
 import { handleProtocol } from './protocol'
 import {
@@ -139,7 +133,15 @@ import {
   wineprefixFAQ
 } from './constants/urls'
 import { legendaryInstalled } from './storeManagers/legendary/constants'
-import { isSnap } from './constants/environment'
+import {
+  isCLIFullscreen,
+  isCLINoGui,
+  isFlatpak,
+  isMac,
+  isSnap,
+  isSteamDeckGameMode,
+  isWindows
+} from './constants/environment'
 
 app.commandLine?.appendSwitch('ozone-platform-hint', 'auto')
