+++ src/backend/shortcuts/ipc_handler.ts
@@ -9,8 +9,8 @@ import {
 } from './nonesteamgame/nonesteamgame'
 import { getInfo } from '../utils'
 import { shortcutFiles } from './shortcuts/shortcuts'
-import { isMac } from 'backend/constants'
 import { notify } from 'backend/dialog/dialog'
+import { isMac } from 'backend/constants/environment'
 
 ipcMain.on('addShortcut', async (event, appName, runner, fromMenu) => {
   gameManagerMap[runner].addShortcuts(appName, fromMenu)
