+++ src/backend/storeManagers/gog/setup.ts
@@ -8,7 +8,6 @@ import {
 } from '../../utils'
 import { GameConfig } from '../../game_config'
 import { logError, logInfo, LogPrefix, logWarning } from '../../logger/logger'
-import { isWindows } from '../../constants'
 import { getWinePath, runWineCommand, verifyWinePrefix } from '../../launcher'
 import { getGameInfo as getGogLibraryGameInfo } from 'backend/storeManagers/gog/library'
 import { readFile } from 'node:fs/promises'
@@ -19,6 +18,7 @@ import {
   GOGv2Manifest
 } from 'common/types/gog'
 import { gogdlConfigPath, gogRedistPath, gogSupportPath } from './constants'
+import { isWindows } from 'backend/constants/environment'
 
 /*
  * Automatially executes command properly according to operating system
