+++ src/backend/storeManagers/legendary/eos_overlay/eos_overlay.ts
@@ -3,7 +3,6 @@ import { existsSync, readFileSync } from 'graceful-fs'
 import { t } from 'i18next'
 import { join } from 'path'
 
-import { isLinux } from 'backend/constants'
 import { logError, LogPrefix, logWarning } from 'backend/logger/logger'
 import { callAbortController } from 'backend/utils/aborthandler/aborthandler'
 import { sendGameStatusUpdate } from 'backend/utils'
@@ -17,6 +16,7 @@ import { Path } from 'backend/schemas'
 import type { Runner } from 'common/types'
 import { toolsPath } from 'backend/constants/paths'
 import { legendaryConfigPath } from '../constants'
+import { isLinux } from 'backend/constants/environment'
 
 const currentVersionPath = () =>
   join(legendaryConfigPath, 'overlay_version.json')
