+++ src/backend/storeManagers/nile/games.ts
@@ -27,7 +27,6 @@ import {
   logFileLocation,
   logInfo
 } from 'backend/logger/logger'
-import { isLinux, isWindows } from 'backend/constants'
 import { GameConfig } from 'backend/game_config'
 import {
   getKnownFixesEnvVariables,
@@ -66,6 +65,7 @@ import { removeNonSteamGame } from 'backend/shortcuts/nonesteamgame/nonesteamgam
 import { sendFrontendMessage } from 'backend/main_window'
 import setup from './setup'
 import { getUmuId } from 'backend/wiki_game_info/umu/utils'
+import { isLinux, isWindows } from 'backend/constants/environment'
 
 export async function getSettings(appName: string): Promise<GameSettings> {
   const gameConfig = GameConfig.get(appName)
