+++ src/backend/storeManagers/storeManagerCommon/games.ts
@@ -1,6 +1,6 @@
 import { GameInfo, GameSettings, Runner } from 'common/types'
 import { GameConfig } from '../../game_config'
-import { isMac, isLinux, icon } from '../../constants'
+import { icon } from '../../constants'
 import {
   appendGamePlayLog,
   appendWinetricksGamePlayLog,
@@ -33,6 +33,7 @@ import {
 import { BrowserWindow, dialog, Menu } from 'electron'
 import { gameManagerMap } from '../index'
 import { sendGameStatusUpdate } from 'backend/utils'
+import { isLinux, isMac } from 'backend/constants/environment'
 
 async function getAppSettings(appName: string): Promise<GameSettings> {
   return (
