+++ src/backend/tools/ipc_handler.ts
@@ -2,10 +2,10 @@ import { gameManagerMap } from 'backend/storeManagers'
 import { ipcMain } from 'electron'
 import { Winetricks, runWineCommandOnGame } from '.'
 import path from 'path'
-import { isWindows } from 'backend/constants'
 import { execAsync, sendGameStatusUpdate } from 'backend/utils'
 import * as GOGLibraryManager from 'backend/storeManagers/gog/library'
 import { sendFrontendMessage } from 'backend/main_window'
+import { isWindows } from 'backend/constants/environment'
 
 ipcMain.handle(
   'runWineCommandForGame',
