+++ src/backend/wiki_game_info/wiki_game_info.ts
@@ -8,8 +8,8 @@ import { logError, logInfo, LogPrefix } from '../logger/logger'
 import { getInfoFromAppleGamingWiki } from './applegamingwiki/utils'
 import { getHowLongToBeat } from './howlongtobeat/utils'
 import { getInfoFromPCGamingWiki } from './pcgamingwiki/utils'
-import { isMac, isLinux } from '../constants'
 import { getUmuId } from './umu/utils'
+import { isLinux, isMac } from 'backend/constants/environment'
 
 export async function getWikiGameInfo(
   title: string,
